onboarding fixes
